requirements.txt
scikit-learn==0.22
pillow==6.2.1
keras==2.3.1
tensorflow-gpu==1.15.0
torch==1.3.1